+++ src/components/index.ts
@@ -6,36 +6,35 @@ export { default as DropdownItem } from './common/dropdown/DropdownItem';
 export { default as DropdownSeparator } from './common/dropdown/DropdownSeparator';
 export { default as DropdownTrigger } from './common/dropdown/DropdownTrigger';
 
+export { default as FollowButton } from './common/FollowButton';
 //modal
 export { default as Modal } from './common/modal/Modal';
+
 export { default as ModalContent } from './common/modal/ModalContent';
-
 export { default as ModalDescription } from './common/modal/ModalDescription';
-export { default as ModalFooter } from './common/modal/ModalFooter';
 
+export { default as ModalFooter } from './common/modal/ModalFooter';
 export { default as ModalHeader } from './common/modal/ModalHeader';
 export { default as ModalTitle } from './common/modal/ModalTitle';
+
 export { default as ModalTrigger } from './common/modal/ModalTrigger';
-
-//layout
 export { default as Header } from './layout/Header';
-export { default as MainLayout } from './layout/MainLayout';
 
+export { default as MainLayout } from './layout/MainLayout';
 export { default as SectionHeader } from './layout/SectionHeader';
 export { default as Sidebar } from './layout/Sidebar';
+
 export { Button, buttonVariants } from './ui/Button';
 
-//ui
 export * from './ui/Dialog';
 
 export * from './ui/DropdownMenu';
-
 export { Input } from './ui/Input';
 export { Loading } from './ui/Loading';
-export { Toaster } from './ui/Sonner';
 
-//card
+export { Toaster } from './ui/Sonner';
 export { default as ConcertCard } from '@/components/common/card/ConcertCard';
+
 export { default as OngoingLiveCard } from '@/components/common/card/OngoingLiveCard';
 
 export { default as CategorySection } from '@/components/common/CategorySection';
