+++ src/types/main.ts
@@ -1,0 +1,11 @@
+export type Artist = {
+  followerText?: string;
+  id: number;
+  imageUrl: string;
+  name: string;
+};
+
+export type CategoryItem = {
+  id: string;
+  label: string;
+};
